test(davinci): keep DOM parity evidence within gates
