+++ crates/vize_s1_to_s2/src/emit/children.rs
@@ -1,5 +1,8 @@
 //! Text and interpolation children.
 
+mod slot_compound;
+
+use slot_compound::emit_slot_compound_parts;
 use vize_davinci::id::NodeId;
 use vize_s0::Span;
 use vize_s2::expr::{ExprRef, JsExpr, OpaqueReason};
@@ -325,31 +328,4 @@ pub(super) fn emit_raw_interpolation_or_refuse(
             expression.span(),
         ))
     }
-}
-
-fn emit_slot_compound_parts(
-    cx: &mut EmitCx<'_>,
-    parts: &[TextPart],
-    span: Span,
-) -> Result<(), EmitError> {
-    if parts.is_empty() {
-        return Err(EmitError::unsupported_at(Reason::EmptyCompoundText, span));
-    }
-    for (i, part) in parts.iter().enumerate() {
-        if i > 0 {
-            cx.buf.push(",");
-            cx.buf.newline();
-        }
-        cx.buf.use_create_text();
-        cx.buf.push(Buf::create_text_alias());
-        cx.buf.push("(");
-        if part.dynamic {
-            emit_to_display_string(cx, part.text.as_str());
-            cx.buf.push(", 1 /* TEXT */");
-        } else {
-            emit_quoted_text(cx, part.text.as_str());
-        }
-        cx.buf.push(")");
-    }
-    Ok(())
 }
