+++ crates/vize_s1_to_s2/tests/davinci_dom_corpus.rs
@@ -10,12 +10,13 @@
 
 #![allow(clippy::disallowed_macros, clippy::disallowed_types)]
 
-use std::{collections::BTreeMap, fs};
+mod davinci_dom_corpus_support;
 
+use davinci_dom_corpus_support::{
+    Report, assert_clean_corpus, assert_empty, compare_sfc_template, compare_sweep,
+    old_lane_skip_is_allowed,
+};
 use vize_atelier_dom::errors::ErrorCode;
-use vize_atelier_sfc::{SfcParseOptions, parse_sfc};
-use vize_s0::Allocator;
-use vize_s1_to_s2::{EmitError, emit_dom_source};
 
 const BATTERY: &[(&str, &str)] = &[
     (
@@ -33,120 +34,6 @@ const BATTERY: &[(&str, &str)] = &[
     (
         "self_closing_non_void_html",
         r#"<template><div /><span class="x" /></template>"#,
-    ),
-];
-
-#[derive(Default)]
-struct Report {
-    files: u64,
-    unreadable_count: u64,
-    parsed: u64,
-    templates: u64,
-    compared: u64,
-    old_error_skips: u64,
-    s2_refusal_count: u64,
-    divergence_count: u64,
-    old_error_codes: Vec<ErrorCode>,
-    old_error_reasons: BTreeMap<String, u64>,
-    unreadable: Vec<String>,
-    old_error_samples: Vec<String>,
-    unexpected_old_error_skips: u64,
-    unexpected_old_error_samples: Vec<String>,
-    s2_refusal_reasons: BTreeMap<&'static str, u64>,
-    s2_refusal_samples: BTreeMap<&'static str, Vec<String>>,
-    s2_refusals: Vec<String>,
-    divergences: Vec<String>,
-}
-
-const OLD_LANE_INVALID_FIXTURES: &[(&str, &[&str])] = &[
-    (
-        "crater/resources/scripts/admin/views/settings/PreferencesSetting.vue",
-        &[
-            "ExtendPoint",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-        ],
-    ),
-    (
-        "dashy/src/components/Widgets/MvgConnection.vue",
-        &["VIfSameKey"],
-    ),
-    (
-        "gogocode/packages/gogocode-vue-playground/packages/vue3/src/components/slots-unification/Comp-out.vue",
-        &["VSlotDuplicateSlotNames"],
-    ),
-    (
-        "habitica/website/client/src/components/modifyInventory.vue",
-        &[
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "InvalidEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-            "MissingEndTag",
-        ],
-    ),
-    (
-        "habitica/website/client/src/components/static/privacy.vue",
-        &["InvalidEndTag"],
-    ),
-    (
-        "heyui/src/components/carousel/carousel.vue",
-        &["VIfSameKey"],
-    ),
-    (
-        "heyui/src/components/table/table.vue",
-        &["VElseNoAdjacentIf"],
-    ),
-    ("tdesign/site/src/pages/design/fonts.vue", &["VIfSameKey"]),
-    (
-        "tdesign/site/src/pages/design/fonts_zh-CN.vue",
-        &["VIfSameKey"],
-    ),
-    (
-        "vue-manage-system/src/views/table/basetable.vue",
-        &["InvalidEndTag"],
-    ),
-    (
-        "vue2-elm/src/page/shop/shop.vue",
-        &["MissingWhitespaceBetweenAttributes"],
-    ),
-    (
-        "vuesax/docs/.vuepress/theme/Home.vue",
-        &["InvalidEndTag", "InvalidEndTag"],
-    ),
-    (
-        "vuesax/docs/.vuepress/theme/homePatreons.vue",
-        &["InvalidEndTag"],
-    ),
-    (
-        "vux/src/components/inline-x-number/index.vue",
-        &["MissingWhitespaceBetweenAttributes"],
-    ),
-    (
-        "vux/src/components/x-number/index.vue",
-        &["MissingWhitespaceBetweenAttributes"],
-    ),
-    (
-        "vux/src/demos/PopupPicker.vue",
-        &["MissingWhitespaceBetweenAttributes"],
     ),
 ];
 
@@ -180,23 +67,7 @@ fn dom_emit_agrees_on_sfc_templates_body() {
         sweep.root.display()
     );
 
-    let mut corpus = Report::default();
-    for file in &sweep.files {
-        let source = match fs::read_to_string(file) {
-            Ok(source) => source,
-            Err(error) => {
-                corpus.unreadable_count += 1;
-                if corpus.unreadable.len() < 20 {
-                    corpus
-                        .unreadable
-                        .push(format!("{}: {error}", file.display()));
-                }
-                continue;
-            }
-        };
-        let context = file.to_string_lossy();
-        compare_sfc_template(context.as_ref(), &source, &mut corpus);
-    }
+    let corpus = compare_sweep(&sweep);
     eprintln!(
         "davinci DOM corpus sweep: files={} unreadable={} parsed={} templates={} compared={} old_error_skips={} s2_refusals={} divergences={}",
         corpus.files,
@@ -225,177 +96,6 @@ fn dom_emit_agrees_on_sfc_templates_body() {
         "a corpus sweep that compares nothing proves nothing"
     );
     assert_clean_corpus(&corpus);
-}
-
-fn compare_sfc_template(name: &str, source: &str, report: &mut Report) {
-    report.files += 1;
-    let Ok(descriptor) = parse_sfc(source, SfcParseOptions::default()) else {
-        return;
-    };
-    report.parsed += 1;
-    let Some(template) = descriptor.template.as_ref() else {
-        return;
-    };
-    report.templates += 1;
-
-    let old_allocator = Allocator::new();
-    let (_, errors, old) = vize_atelier_dom::compile_template(&old_allocator, &template.content);
-    let blocking_errors: Vec<_> = errors
-        .iter()
-        .filter(|error| !error.is_recoverable())
-        .collect();
-    if !blocking_errors.is_empty() {
-        report.old_error_skips += 1;
-        for error in &blocking_errors {
-            *report
-                .old_error_reasons
-                .entry(format!("{:?}", error.code))
-                .or_default() += 1;
-        }
-        if report.old_error_codes.len() < 20 {
-            report
-                .old_error_codes
-                .extend(blocking_errors.iter().map(|error| error.code));
-        }
-        if report.old_error_samples.len() < 20 {
-            report.old_error_samples.push(format!(
-                "{name}: {} old-lane blocking errors: {blocking_errors:?}",
-                blocking_errors.len()
-            ));
-        }
-        let actual_codes = blocking_errors
-            .iter()
-            .map(|error| format!("{:?}", error.code))
-            .collect::<Vec<_>>();
-        if !old_lane_skip_is_allowed(name, &actual_codes) {
-            report.unexpected_old_error_skips += 1;
-            if report.unexpected_old_error_samples.len() < 20 {
-                report.unexpected_old_error_samples.push(format!(
-                    "{name}: unexpected old-lane blocking errors: {blocking_errors:?}"
-                ));
-            }
-        }
-        return;
-    }
-    let old = format!("{}\n{}", old.preamble, old.code);
-
-    let new_allocator = Allocator::new();
-    let new = match emit_dom_source(&new_allocator, &template.content) {
-        Ok(emit) => emit.assembled(),
-        Err(error) => {
-            report.s2_refusal_count += 1;
-            let reason = refusal_reason(&error);
-            *report.s2_refusal_reasons.entry(reason).or_default() += 1;
-            let samples = report.s2_refusal_samples.entry(reason).or_default();
-            if samples.len() < 5 {
-                samples.push(format!("{name}: {error:?}"));
-            }
-            if report.s2_refusals.len() < 20 {
-                report.s2_refusals.push(format!("{name}: {error:?}"));
-            }
-            return;
-        }
-    };
-
-    report.compared += 1;
-    if old != new {
-        report.divergence_count += 1;
-        if report.divergences.len() < 20 {
-            report.divergences.push(format!(
-                "{name}: old_len={} new_len={} first_diff={} old_window={} new_window={}",
-                old.len(),
-                new.len(),
-                first_diff(&old, &new),
-                mismatch_window(&old, &new),
-                mismatch_window(&new, &old)
-            ));
-        }
-    }
-}
-
-fn refusal_reason(error: &EmitError) -> &'static str {
-    error.reason().map_or("diagnostics", |reason| reason.code())
-}
-
-fn preview(source: &str) -> String {
-    source
-        .lines()
-        .take(4)
-        .collect::<Vec<_>>()
-        .join("\\n")
-        .chars()
-        .take(320)
-        .collect()
-}
-
-fn first_diff(left: &str, right: &str) -> usize {
-    left.as_bytes()
-        .iter()
-        .zip(right.as_bytes())
-        .position(|(left, right)| left != right)
-        .unwrap_or_else(|| left.len().min(right.len()))
-}
-
-fn mismatch_window(source: &str, other: &str) -> String {
-    let diff = first_diff(source, other);
-    let start = source[..diff]
-        .char_indices()
-        .rev()
-        .nth(80)
-        .map_or(0, |(index, _)| index);
-    let end = source[diff..]
-        .char_indices()
-        .nth(180)
-        .map_or(source.len(), |(index, _)| diff + index);
-    preview(&source[start..end])
-}
-
-fn assert_empty(label: &str, values: &[String]) {
-    assert!(values.is_empty(), "{label}:\n{}", values.join("\n"));
-}
-
-fn assert_clean_corpus(report: &Report) {
-    assert!(
-        report.unreadable_count == 0
-            && report.unexpected_old_error_skips == 0
-            && report.s2_refusal_count == 0
-            && report.divergence_count == 0,
-        "corpus unreadable files ({}):\n{}\n\ncorpus old-lane error skips ({}) by reason {:?}:\n{}\n\nunexpected old-lane error skips ({}):\n{}\n\ncorpus S2 refusals ({}) by reason {:?}:\n{}\n\ncorpus divergences ({}):\n{}",
-        report.unreadable_count,
-        report.unreadable.join("\n"),
-        report.old_error_skips,
-        report.old_error_reasons,
-        report.old_error_samples.join("\n"),
-        report.unexpected_old_error_skips,
-        report.unexpected_old_error_samples.join("\n"),
-        report.s2_refusal_count,
-        report.s2_refusal_reasons,
-        report.s2_refusals.join("\n"),
-        report.divergence_count,
-        report.divergences.join("\n"),
-    );
-}
-
-fn old_lane_skip_is_allowed(name: &str, actual_codes: &[String]) -> bool {
-    let Some(path) = fixture_path(name) else {
-        return false;
-    };
-    let Some((_, expected_codes)) = OLD_LANE_INVALID_FIXTURES
-        .iter()
-        .find(|(allowed_path, _)| *allowed_path == path)
-    else {
-        return false;
-    };
-    let mut actual = actual_codes.iter().map(String::as_str).collect::<Vec<_>>();
-    actual.sort_unstable();
-    let mut expected = expected_codes.to_vec();
-    expected.sort_unstable();
-    actual == expected
-}
-
-fn fixture_path(name: &str) -> Option<&str> {
-    name.split_once("tests/_fixtures/_git/")
-        .map(|(_, path)| path)
 }
 
 #[test]
