+++ davinci-road/plan/storage-inventory.tsv
@@ -42,8 +42,9 @@ s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/component/preamble.rs	1	3	0	0	0	0	0	
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/create_slots.rs	1	6	1	6	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/directive.rs	1	6	0	0	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/filter.rs	0	0	1	1	0	0	0	0
-s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/hoist.rs	1	4	1	12	0	0	0	0
+s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/hoist.rs	1	1	1	11	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/hoist/cached_props.rs	0	0	1	1	0	0	0	0
+s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/hoist/props.rs	1	3	1	3	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/entity.rs	0	0	1	3	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/js.rs	0	0	1	6	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/js_comment.rs	0	0	1	3	0	0	0	0
@@ -57,7 +58,7 @@ s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/props.rs	1	2	1	4	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/props_bind.rs	1	5	1	2	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/props_object.rs	1	5	0	0	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/props_object_merge.rs	0	0	1	1	0	0	0	0
-s1_to_s2	-	crates/vize_s1_to_s2/src/emit/props_static.rs	0	0	1	6	0	0	0	0
+s1_to_s2	-	crates/vize_s1_to_s2/src/emit/props_static.rs	0	0	1	10	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/slots.rs	1	5	1	6	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/tpl.rs	0	0	1	3	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/vfor.rs	0	0	1	4	0	0	0	0
